requirements.txt
pyyaml
beautifulsoup4
requests
Pillow
reportlab